fix([DST-1407]): keep parent Drawer mounted when nested trigger opens
A `<Drawer.Trigger>` rendered inside an open Drawer used to dismiss the
parent on click, which unmounted the nested trigger and tore down the new
Drawer with it. Detect the nested case via a new `DrawerNestingContext`
and skip registering with the single-slot registry; the nested Drawer
now opens over its parent. "One at a time" continues to apply between
siblings.

Refactor: replace the renderless `<DrawerCoordinator />` component with
a `useDrawerCoordination` hook called inside `Drawer`. Calling at this
level lets the hook read the *ancestor's* nesting context before this
Drawer publishes its own provider for descendants.
