+++ packages/components/src/Drawer/Drawer.tsx
@@ -11,13 +11,13 @@ import { useLandmark } from '@react-aria/landmark';
 import { cn, useClassNames, useSmallScreen } from '@marigold/system';
 import { CloseButton } from '../CloseButton/CloseButton';
 import { intlMessages } from '../intl/messages';
-import { DrawerContext } from './Context';
+import { DrawerContext, DrawerNestingContext } from './Context';
 import { DrawerActions } from './DrawerActions';
 import { DrawerContent } from './DrawerContent';
-import { DrawerCoordinator } from './DrawerCoordinator';
 import { DrawerModal } from './DrawerModal';
 import { DrawerTitle } from './DrawerTitle';
 import { DrawerTrigger } from './DrawerTrigger';
+import { useDrawerCoordination } from './useDrawerCoordination';
 
 // Props
 // ---------------
@@ -79,6 +79,10 @@ export const Drawer = ({
 
   const ctx = use(OverlayTriggerStateContext);
 
+  // Called here (not in a child) so it reads the ancestor's `DrawerNestingContext`,
+  // not the provider this component publishes below.
+  useDrawerCoordination();
+
   /**
    * On smaller screens the we render a modal dialog instead of a non-modal drawer
    * and need to adjust the role and props accordingly.
@@ -94,30 +98,31 @@ export const Drawer = ({
       keyboardDismissable={keyboardDismissable}
       data-testid="drawer-modal"
     >
-      <DrawerContext value={{ variant, size }}>
-        <Dialog
-          {...props}
-          // Override RAC here so we can set an appropriate role
-          {...(landmarkProps as any)}
-          className={cn(
-            'h-(--visual-viewport-height) outline-none',
-            // Use single quotes, in some enviroments the class is not correctly applied otherwise
-            "grid [grid-template-areas:'title'_'content'_'actions']",
-            classNames.container
-          )}
-        >
-          <DrawerCoordinator />
-          {closeButton && (
-            <CloseButton
-              aria-label={stringFormatter.format('dismissDrawer')}
-              style={{ '--i': 0 } as CSSProperties}
-              className={cn('z-80', classNames.closeButton)}
-              onPress={ctx?.close}
-            />
-          )}
-          {children}
-        </Dialog>
-      </DrawerContext>
+      <DrawerNestingContext value={true}>
+        <DrawerContext value={{ variant, size }}>
+          <Dialog
+            {...props}
+            // Override RAC here so we can set an appropriate role
+            {...(landmarkProps as any)}
+            className={cn(
+              'h-(--visual-viewport-height) outline-none',
+              // Use single quotes, in some enviroments the class is not correctly applied otherwise
+              "grid [grid-template-areas:'title'_'content'_'actions']",
+              classNames.container
+            )}
+          >
+            {closeButton && (
+              <CloseButton
+                aria-label={stringFormatter.format('dismissDrawer')}
+                style={{ '--i': 0 } as CSSProperties}
+                className={cn('z-80', classNames.closeButton)}
+                onPress={ctx?.close}
+              />
+            )}
+            {children}
+          </Dialog>
+        </DrawerContext>
+      </DrawerNestingContext>
     </DrawerModal>
   );
 };
